Moved Shapes into namespace
